fix(rust): satisfy Rust 1.85 MSRV clippy lints
CI runs clippy on both stable and the 1.85 MSRV toolchain. The 1.85
clippy flags three lints that newer stable no longer reports:
- unnecessary_cast on an already-i32 iterations field (harness.rs)
- format_collect in two nonce generators (parsers/prompty.rs,
  renderers/common.rs)

Remove the no-op cast and rebuild nonce strings with fold + write!,
which is behavior-identical and clean under both toolchains.

+++ runtime/rust/prompty/src/harness.rs
@@ -1188,7 +1188,7 @@ where
             turn_id: request.turn_id,
             status,
             output,
-            iterations: result.commit.iterations as i32,
+            iterations: result.commit.iterations,
             tool_results,
             checkpoints,
         })
